API preparation for the 0.4.4 release.
This is just an update on the "API" section with all the stuff as an
uppdate on the "Android App" and "Web UI" should happen in the next
couple of weeks.
== Change Log ==
* Changed the RestrictedAPICore to be a PHP Object with private
variables and methods (as opposed to a set of global variables and
functions)
* Changed how the device permission system works (no longer just premise
based where you had to give the user access to everything in the premise
or nothing) to a room based premise system (so that you can give a user
permission to access just the devices in the rooms specified)(This will
be changed to a different system in the future)
* Added a new permission system for the WatchInputs users so that they
won't be affected by the normal device permission system.
* Added a HubTelnet API with some common functions in it to allow for
telnetting into WatchInputs and getting it to turn on Zigbee join mode
to allow compatible zigbee devices to be added to the iOmy system.
* Relocated the weather library to a more appropriate location
* Added a new mode to the OnvifThumbnail API to allow just fetching the
most recent cached image off the server rather than creating a new
thumbnail so that an old image can be displayed while waiting for the
new one to be created.
* Added APIs to create new users as well as editing their permissions.