Updated the pages for creating or editing rooms and premises in preparation for 0.4.4 release
@BrentJarmaine - Updated the pages for creating or editing rooms and
premises in preparation for 0.4.4 release
== Change Log ==
In RoomAdd.controller.js and RoomEdit.controller.js:
* Commented out references to the floor the room belongs to for now.
* Changed an error message that actually reflects the room page rather
than any premise page.
* Kept the submit button on each of the forms disabled until the request
is complete and either the user is taken back to the previous page and
goes back in to one of the forms, or, if unsuccessful, after the error
message is closed.

In PremiseEditAddress.controller.js and PremiseEditInfo.controller.js:
* Removed the red asterix indicating a required field from labels above
the new select boxes.
* Switch from using combo boxes to select boxes so that there will
always be a value to get from the widget without the chance of the user
entering something he/she shouldn't in the input part of the select box.
Changed which functions to use for grabbing information from selected
options.