Updated UI functionality in preparation for the 0.4.4 release
@BrentJarmaine
== Change Log ==

* Unassigned Devices page is now obsolete.
* Changed the "Extras Menu" to "Action Menu" and now has a plus button
in a circle as the icon.
* Various styling changes
* New functions for trying to deal with some of the UI refresh issues.

In DeviceList.controller.js
* Restructured code to make it easier to maintain.

In EditLink.controller.js and EditThing.controller.js:
* Have the submit button remain disabled until either the user goes back
to it after being taken to the previous page after a successful change,
or after the error dialog is shown.
* In EditThing, changed the property names of the thing object to use
the original names that were defined when the IOMy.common.ThingList
global variable was populated.

In LinkAdd.controller.js:
* Created the functionality to assign a newly created link to a given
room.
* Created the function to change aspects of the link form based on
currently selected link type.
* Restructured portions of code.
* New fields for selecting the premise and one of its rooms.

In app.css and experimental_brent.css
* New styles for the UI.
* Straightened the format of the CSS code.

In functions.js:
* New function to retrieve the ID of the premise that a given hub is
connected to.
* Functions to try and deal with the UI refresh problem.