watch inputs webapiclientlib: Changed zigbee link display name to empty
The user modifiable display name that watch inputs was using for zigbee
devices when uploading as a link to the database wasn't unique per
device so multiple zigbee devices of the same brand and model would get
the same display name which was confusing for users.  Changed to an
empty string so the user interface can come up with a unique name for
each device until the user changes it to something they prefer.