watch inputs: Implemented processing and reloading of time rules
timeruleslib now processes the time rules and set flags when need to
either turn on or off a device and you can also reload the time rules
file with a telnet command.  Also removed some unused variables and
changed some uses of ifaceptr to auto const so the compiler can better
alert for invalid uses of them.