iomyapp: First Run Wizard now keeps track of where the user is up to
Added tracking via Android Preferences where the user is currently up to
while in the first run wizard so if they fully close the app, when they
reload it will re-open where the wizard left off.  There are still a few
sub-stages that will fail if the app is exited while they are running,
but it is better than it was before.