watch_inputs: New netnput function and bugfix in netputs
commonserverlib:

Bugfix: netputs wasn't incrementing pos by the sent len after sending.
This would cause truncated output if send failed to send the entire
string in the first call.
Added a function: netnput that is similar to netputs but the caller can
specify the length of the string as an argument