iomyapp: Improved user info during MySQL loading, improved upgrade scripts
This comment describes the previous commit as well.
Improved user info during MySQL loading and better handling and info for
if MySQL fails to load.  Added debug logging of mysqlcheck program for
advanced diagnostics.  Disabled checking of database when in demo mode
as it needs the root password which isn't stored for demo database.
MySQL check if running now uses an alternative check that includes
checking if the socket exists as that often takes a while to get created
which is the real indicator of when MySQL is ready to service requests.

Improved the upgrade scripts and they are now running incrementally from
the user's previous version up until the current version so we only need
an upgrade script for each incremental version upgrade instead of a
script for every combination of version differences.