iomyapp: Moved demo data into a package that the app downloads
Demo data is very large so packaging in the app was making the apk very
large.  This moves the demo data to an offsite location and the app now
downloads the demo data when the user selects to enable demo mode in the
install wizard.  Also updated the extractor to support multiple files
and set demo mode in the install wizard to be unticked by default as it
can take a while to download and is easy to miss that the option is
available.
Added some functions to Application class to get the version name and
code so the correct demo data for the app version can be downloaded.