UI - Major changes to the Core Variables
== Change Log ==
* Swapped the main refresh core variables function back to an older
function since the newer one seems to be used for loading the non-core
variables first.
* Made a seperate function to load the optional variables (non-core)
until they have seperate functions for each to only load them as needed
rather than as soon as it logs in.
* Added a RefreshUserInfo function to the UI common modules.
* Changes to the Odata API and PHP API ajax functions to have them
silently abort if it becomes orphaned (eg. Another API has flagged that
the User is no longer logged in).
* Added a preset config to the UI common module to manage what occurs
after the first refresh core variables happens in the 2 seperate places
in the Login.controller (when the user either logs in and the other is
the UI sees that the user is already logged in).