watch_inputs: nativeseriallib: Fixed: filename is not unique enough
nativeseriallib was using just the filename to identify the slot to
access, but often if a device is removed and re-added quickly, there
will be two active slots with the same filename.  Then the newly added
entry would be sometimes be removed instead of the removed entry as the
new entry would match first, and then cause a crash from accessing freed
data.

Fixed by adding the slot number to the uniqueid and matching on the
uniqueid instead of just the filename.