watch_inputs: Added Bluetooth CSRMesh library
Added support for CSRMesh over Bluetooth on Android to connect to
CSRMesh Light Bulbs.

In webapiclientlib, Comm and Link queues are now derived from a base
comm and link queues to reduce duplication in the api for different
types of links and polymorphism is used to add the correct values based
on the object type.  Also added a Thing queue for adding Things separate
from adding Links.

In dblib and mysqllib, Added database query functions needed by the new
CSRMesh library.

Added *.o to gitignore as gcc will generate those if it is run on a c or
c++ file.

Added uses BLUETOOTH permission to android watch inputs manifest.

In Android MysqlLib, added getInstance() function so Context object doesn't
have to be static.