watch_inputs: cmdserverlib: Added support for username & password
Added support for configuring a username and password for cmdserverlib
in the config file.  If a username or password isn't configured then
login will be allowed with no authentication required.  Also, tells
telnet clients to hide screen output during password input.

Added telnet initialisation sequence at the beginning of the session so
Telnet clients can be detected and configured correctly