watch_inputs: commonserverlib: netgetc_with_quitpipe: Fix buffer overflow
In commonserverlib's netgetc_with_quitpipe function, the received and
pos variables weren't getting reset to 0 at the start of the reading
loop so if the loop aborted early before getting to the recv operation,
received and pos would still be set to the values from the previous
call when returning the next character.  If received was still > 0, the
last operation before returning would then be to increment pos which
would eventually cause the returned value to go past the end of the
received buffer contents if continuously called.

Also lerror was still being set to errno even if recv returned success.
We now set lerrno to 0 in case errno is set to something else.