watch_inputs: ruleslib: Added handling of Philips Hue
Added handling of rules to turn a Philips Hue device on or off.

Philips Hue also requires setting on/off via the webapi so added
handling of that and also added the ability to trigger a webapi request
immediately instead of waiting for the normal timer to trigger the next
request.