watch_inputs: cameralib: Added various test telnet commands
Added various test telnet commands to cameralib to enable the ability to
do full testing without needing the webapi including the ability to
remove a camera object and camera stream object and start and stop a
camera stream, and set camera stream properties.

The forked child now redirects stdin, stdout, and stderr to /dev/null as
without that it can't exit correctly while the parent is running.

Added ability to auto restart or auto delete a camera stream when it
exits.